Improve code reuse
