+++ react-app/utils/service-worker/generate.js
@@ -1,12 +1,16 @@
 const fs = require('fs');
-const { join } = require('path');
+const path = require('path');
 const { injectManifest } = require('workbox-build');
+const getEntrypoints = require('../entrypoints');
 
-const BUILD_DIR = join(__dirname, '..', 'build');
+const BUILD_DIR = path.join(__dirname, '..', '..', 'build');
 
-const ASSET_MANIFEST_PATH = join(BUILD_DIR, 'asset-manifest.json');
-const SW_PATH = join(BUILD_DIR, 'service-worker.js');
-const SW_TEMPLATE_PATH = join(__dirname, 'template.js');
+const ASSET_MANIFEST_PATH = path.join(BUILD_DIR, 'asset-manifest.json');
+const SW_PATH = path.join(BUILD_DIR, 'service-worker.js');
+const SW_TEMPLATE_PATH = path.join(__dirname, 'template.js');
+
+const APP_ROOT = '/content/wknd-events/react';
+const APP_START_PAGE = `${APP_ROOT}/home.html`;
 
 /**
  * Deletes a previously generated service worker file (if it exists)
@@ -20,30 +24,6 @@ const removeExistingWorker = swPath => {
   } else {
     console.log(`No existing service worker at ${swPath}`);
   }
-};
-
-/**
- * Returns all entrypoint chunks (JS and CSS) of the React app. These chunks
- * will not need to be precached because they're already requested from the HTML
- * file)
- *
- * @param {string} assetManifestPath: Path to the asset manifest file from which
- * the entrypoint files can be read
- */
-const getEntrypoints = assetManifestPath => {
-  if (!fs.existsSync(assetManifestPath)) {
-    throw Error(
-      `Cannot determine entrypoints: No asset manifest found at path ${assetManifestPath}`
-    );
-  }
-  const manifest = fs.readFileSync(assetManifestPath, { encoding: 'utf8' });
-  const manifestContent = JSON.parse(manifest);
-  if (!('entrypoints' in manifestContent)) {
-    throw Error(
-      `Cannot determine entrypoints: Missing "entrypoints" key in ${assetManifestPath}`
-    );
-  }
-  return manifestContent.entrypoints;
 };
 
 /**
@@ -68,7 +48,7 @@ const main = async () => {
       originalManifest => ({
         manifest: originalManifest.map(entry =>
           entry.url.endsWith('index.html')
-            ? { ...entry, url: '/content/wknd-events/react/home.html' }
+            ? { ...entry, url: APP_START_PAGE }
             : entry
         ),
         warnings: []
