+++ react-app/utils/service-worker/template.js
